llm-wiki-webapp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
anthropic==0.25.0
python-dateutil==2.9.0
python-multipart==0.0.9
